[REFACTOR] qna 도메인 리펙토링 (#191)

+++ src/main/java/com/programmers/smrtstore/domain/qna/application/ProductQuestionService.java
@@ -1,10 +1,8 @@
 package com.programmers.smrtstore.domain.qna.application;
 
 import com.programmers.smrtstore.core.properties.ErrorCode;
-import com.programmers.smrtstore.domain.qna.domain.entity.ProductAnswer;
 import com.programmers.smrtstore.domain.qna.domain.entity.ProductQuestion;
 import com.programmers.smrtstore.domain.qna.exception.QnAException;
-import com.programmers.smrtstore.domain.qna.infrastructure.ProductAnswerRepository;
 import com.programmers.smrtstore.domain.qna.infrastructure.ProductQuestionRepository;
 import com.programmers.smrtstore.domain.qna.presentation.dto.req.*;
 import com.programmers.smrtstore.domain.qna.presentation.dto.res.*;
@@ -20,15 +18,13 @@ import java.util.List;
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class ProductQnAService {
-    private final ProductAnswerRepository productAnswerRepository;
-
+public class ProductQuestionService {
     private final ProductQuestionRepository questionRepository;
     private final UserJpaRepository userJpaRepository;
 
     public CreateQuestionResponse createQuestion(Long userId, CreateQuestionRequest request) {
-        checkUserValid(userId, request.getUserId());
         checkUserExist(userId);
+        checkUserValid(userId, request.getUserId());
         ProductQuestion productQuestion = ProductQuestion.builder()
                 .userId(request.getUserId())
                 .productId(request.getProductId())
@@ -40,21 +36,24 @@ public class ProductQnAService {
 
     @Transactional(readOnly = true)
     public List<QuestionResponse> findByUserId(Long userId, FindQuestionRequest request) {
-        checkUserValid(userId, request.getUserId());
         checkUserExist(userId);
+        checkUserValid(userId, request.getUserId());
         return questionRepository.findByUserId(userId);
     }
 
     public UpdateQuestionResponse updateQuestion(Long userId, UpdateQuestionRequest request) {
         checkUserExist(userId);
         ProductQuestion productQuestion = questionRepository.findById(request.getId()).orElseThrow(() -> new QnAException(ErrorCode.QUESTION_NOT_FOUND));
-        productQuestion.updateContent(request.getContent());
+        productQuestion.updateContent(userId, request.getContent());
         return UpdateQuestionResponse.of(productQuestion);
     }
 
     public DeleteQuestionResponse deleteQuestion(Long userId, Long deleteId) {
         checkUserExist(userId);
         ProductQuestion deleteProductQuestion = questionRepository.findById(deleteId).orElseThrow(() -> new QnAException(ErrorCode.QUESTION_NOT_FOUND));
+        if (!deleteProductQuestion.checkUserEquals(userId)) {
+            throw new QnAException(ErrorCode.QUESTION_ACCESS_DENIED);
+        }
         questionRepository.delete(deleteProductQuestion);
         return DeleteQuestionResponse.of(deleteProductQuestion);
     }
@@ -64,41 +63,14 @@ public class ProductQnAService {
         return questionRepository.findByProductId(productId);
     }
 
-    public AnswerResponse addAnswer(Long userId, CreateAnswerRequest request) {
-        checkUserExist(userId);
-        Long questionId = request.getQuestionId();
-        String content = request.getContent();
-        ProductQuestion productQuestion = questionRepository.findById(questionId).orElseThrow(() -> new QnAException(ErrorCode.QUESTION_NOT_FOUND));
-        ProductAnswer productAnswer = ProductAnswer.builder()
-                .productQuestion(productQuestion)
-                .content(content)
-                .build();
-        return AnswerResponse.of(productAnswer);
-    }
 
-    @Transactional(readOnly = true)
-    public List<AnswerResponse> getAnswersByQuestionId(Long userId, Long questionId) {
-        ProductQuestion productQuestion = questionRepository.findById(questionId).orElseThrow();
-        return productQuestion.getProductAnswerList()
-                .stream()
-                .map(AnswerResponse::of)
-                .toList();
-    }
-
-    public UpdateAnswerResponse updateAnswer(Long userId, UpdateAnswerRequest request) {
-        checkUserExist(userId);
-        ProductAnswer productAnswer = productAnswerRepository.findById(request.getId()).orElseThrow(() -> new QnAException(ErrorCode.ANSWER_NOT_FOUND));
-        productAnswer.updateContent(request.getContent());
-        return UpdateAnswerResponse.of(productAnswer);
-    }
-
-    void checkUserValid(Long tokenUserId, Long requestUserId) {
+    private void checkUserValid(Long tokenUserId, Long requestUserId) {
         if(!tokenUserId.equals(requestUserId)) {
             throw new QnAException(ErrorCode.USER_NOT_FOUND);
         }
     }
 
-    void checkUserExist(Long userId) {
+    private void checkUserExist(Long userId) {
         userJpaRepository.findById(userId).orElseThrow(() -> new QnAException(ErrorCode.USER_NOT_FOUND));
     }
 }
